bugfix/tweak: Исправление недочетов byond 516 + возможность сохранять настройки чата в файл
