+++ code/modules/tgui/tgui_panel/tgui_panel.dm
@@ -60,6 +60,7 @@
 	// window.send_asset(get_asset_datum(/datum/asset/spritesheet/emoji))
 	request_telemetry()
 	addtimer(CALLBACK(src, PROC_REF(on_initialize_timed_out)), 5 SECONDS)
+	window.send_message("testTelemetryCommand")
 
 /**
  * private
