+++ tgui/package.json
@@ -21,6 +21,7 @@
     "@types/node": "^20.14.11",
     "@types/webpack": "^5.28.5",
     "@types/webpack-env": "^1.18.5",
+    "@types/wicg-file-system-access": "^2023.10.5",
     "@typescript-eslint/parser": "^8.0.0-alpha",
     "babel-jest": "^29.7.0",
     "babel-loader": "^8.3.0",
