+++ tgui/packages/tgui-panel/chat/types.ts
@@ -1,0 +1,9 @@
+export type Page = {
+  isMain: boolean;
+  id: string;
+  name: string;
+  acceptedTypes: Record<string, boolean>;
+  unreadCount: number;
+  hideUnreadCount: boolean;
+  createdAt: number;
+};
