+++ tgui/packages/tgui-panel/settings/SettingsPanel.tsx
@@ -1,0 +1,53 @@
+/**
+ * @file
+ * @copyright 2020 Aleksej Komarov
+ * @license MIT
+ */
+
+import { useDispatch, useSelector } from 'common/redux';
+import { Section, Stack, Tabs } from 'tgui/components';
+
+import { ChatPageSettings } from '../chat';
+import { changeSettingsTab } from './actions';
+import { SETTINGS_TABS } from './constants';
+import { selectActiveTab } from './selectors';
+import { SettingsGeneral } from './SettingsGeneral';
+import { TextHighlightSettings } from './TextHighlight';
+import { SettingsStatPanel } from './SettingsStatPanel';
+
+export const SettingsPanel = (props, context) => {
+  const activeTab = useSelector(context, selectActiveTab);
+  const dispatch = useDispatch(context);
+
+  return (
+    <Stack fill>
+      <Stack.Item>
+        <Section fitted fill minHeight="8em">
+          <Tabs vertical>
+            {SETTINGS_TABS.map((tab) => (
+              <Tabs.Tab
+                key={tab.id}
+                selected={tab.id === activeTab}
+                onClick={() =>
+                  dispatch(
+                    changeSettingsTab({
+                      tabId: tab.id,
+                    })
+                  )
+                }
+              >
+                {tab.name}
+              </Tabs.Tab>
+            ))}
+          </Tabs>
+        </Section>
+      </Stack.Item>
+      <Stack.Item grow basis={0}>
+        {activeTab === 'general' && <SettingsGeneral />}
+        {activeTab === 'chatPage' && <ChatPageSettings />}
+        {activeTab === 'textHighlight' && <TextHighlightSettings />}
+        {activeTab === 'statPanel' && <SettingsStatPanel />}
+      </Stack.Item>
+    </Stack>
+  );
+};
