+++ tgui/packages/tgui-panel/settings/middleware.ts
@@ -1,0 +1,165 @@
+/**
+ * @file
+ * @copyright 2020 Aleksej Komarov
+ * @license MIT
+ */
+
+import { storage } from 'common/storage';
+
+import { setClientTheme } from '../themes';
+import {
+  addHighlightSetting,
+  loadSettings,
+  removeHighlightSetting,
+  updateHighlightSetting,
+  updateSettings,
+  exportSettings,
+  importSettings,
+} from './actions';
+import { FONTS_DISABLED } from './constants';
+import { setDisplayScaling } from './scaling';
+import { selectSettings } from './selectors';
+import { exportChatSettings } from './settingsImExport';
+
+let setStatFontFamilyTimer: NodeJS.Timeout;
+let statFontTimer: NodeJS.Timeout;
+let statTabsTimer: NodeJS.Timeout;
+let overrideRule: HTMLStyleElement;
+let overrideFontFamily: string | undefined;
+let overrideFontSize: string;
+
+/** Updates the global CSS rule to override the font family and size. */
+const updateGlobalOverrideRule = () => {
+  let fontFamily = '';
+
+  if (overrideFontFamily !== undefined) {
+    fontFamily = `font-family: ${overrideFontFamily} !important;`;
+  }
+
+  const constructedRule = `body * :not(.Icon) {
+    ${fontFamily}
+  }`;
+
+  if (overrideRule === undefined) {
+    overrideRule = document.createElement('style');
+    document.querySelector('head')!.append(overrideRule);
+  }
+
+  // no other way to force a CSS refresh other than to update its innerText
+  overrideRule.innerText = constructedRule;
+
+  document.body.style.setProperty('font-size', overrideFontSize);
+};
+
+const setGlobalFontSize = (
+  fontSize: string,
+  statFontSize: string,
+  statLinked: boolean
+) => {
+  overrideFontSize = `${fontSize}px`;
+  // Used solution from theme.ts
+  clearInterval(statFontTimer);
+  Byond.command(
+    `.output statbrowser:set_font_size ${statLinked ? fontSize : statFontSize}px`
+  );
+  statFontTimer = setTimeout(() => {
+    Byond.command(
+      `.output statbrowser:set_font_size ${statLinked ? fontSize : statFontSize}px`
+    );
+  }, 1500);
+};
+
+const setGlobalFontFamily = (
+  fontFamily: string,
+  statFontFamily: string,
+  statLinked: boolean
+) => {
+  overrideFontFamily = fontFamily === FONTS_DISABLED ? undefined : fontFamily;
+
+  clearInterval(setStatFontFamilyTimer);
+  Byond.command(
+    `.output statbrowser:set_font_style ${statLinked ? fontFamily : statFontFamily}`
+  );
+  setStatFontFamilyTimer = setTimeout(() => {
+    Byond.command(
+      `.output statbrowser:set_font_style ${statLinked ? fontFamily : statFontFamily}`
+    );
+  }, 1500);
+};
+
+const setStatTabsStyle = (style: string) => {
+  // Well... another timer copy-paste
+  clearInterval(statTabsTimer);
+  Byond.command(`.output statbrowser:set_tabs_style ${style}`);
+  statTabsTimer = setTimeout(() => {
+    Byond.command(`.output statbrowser:set_tabs_style ${style}`);
+  }, 1500);
+};
+
+export const settingsMiddleware = (store) => {
+  let initialized = false;
+
+  return (next) => (action) => {
+    const { type, payload } = action;
+
+    if (!initialized) {
+      initialized = true;
+      setDisplayScaling();
+      storage.get('panel-settings').then((settings) => {
+        store.dispatch(loadSettings(settings));
+      });
+    }
+    if (type === exportSettings.type) {
+      const state = store.getState();
+      const settings = selectSettings(state);
+      exportChatSettings(settings, state.chat.pageById);
+      return;
+    }
+    if (
+      type !== updateSettings.type &&
+      type !== loadSettings.type &&
+      type !== addHighlightSetting.type &&
+      type !== removeHighlightSetting.type &&
+      type !== updateHighlightSetting.type &&
+      type !== importSettings.type
+    ) {
+      return next(action);
+    }
+
+    // Set client theme
+    const theme = payload?.theme;
+    if (theme) {
+      setClientTheme(theme);
+    }
+
+    // Pass action to get an updated state
+    next(action);
+
+    const settings = selectSettings(store.getState());
+
+    if (importSettings.type) {
+      setClientTheme(settings.theme);
+    }
+
+    // Update stat panel personal settings
+    setStatTabsStyle(settings.statTabsStyle);
+
+    // Update global UI font size
+    setGlobalFontSize(
+      settings.fontSize,
+      settings.statFontSize,
+      settings.statLinked
+    );
+    setGlobalFontFamily(
+      settings.fontFamily,
+      settings.statFontFamily,
+      settings.statLinked
+    );
+    updateGlobalOverrideRule();
+
+    // Save settings to the web storage
+    storage.set('panel-settings', settings);
+
+    return;
+  };
+};
